@@ -31,31 +31,32 @@
   "homepage": "https://github.com/Edouardbozon/codamit#readme",
   "dependencies": {
     "@akanass/rx-http-request": "^3.1.0",
-    "@marblejs/core": "^1.2.1",
-    "@marblejs/middleware-body": "^1.2.1",
-    "@marblejs/middleware-joi": "^1.2.1",
-    "@marblejs/middleware-jwt": "^1.2.1",
-    "@marblejs/middleware-logger": "^1.2.1",
+    "@marblejs/core": "^2.0.0-rc.1",
+    "@marblejs/middleware-body": "^2.0.0-rc.1",
+    "@marblejs/middleware-io": "^2.0.0-rc.1",
+    "@marblejs/middleware-joi": "^2.0.0-rc.1",
+    "@marblejs/middleware-jwt": "^2.0.0-rc.1",
+    "@marblejs/middleware-logger": "^2.0.0-rc.1",
     "@sentry/browser": "^4.4.2",
-    "@webcomponents/webcomponentsjs": "^2.2.1",
+    "@webcomponents/webcomponentsjs": "^2.2.7",
     "anchorme": "^1.1.2",
-    "bcrypt": "^3.0.3",
-    "bulma": "^0.7.2",
-    "chalk": "^2.4.1",
+    "bcrypt": "^3.0.4",
+    "bulma": "^0.7.4",
+    "chalk": "^2.4.2",
     "date-fns": "^1.30.1",
     "dotenv": "^6.2.0",
-    "dotenv-webpack": "^1.6.0",
+    "dotenv-webpack": "^1.7.0",
     "express": "^4.16.4",
     "flyd": "^0.2.8",
-    "lit-element": "2.0.0-rc.3",
-    "lit-html": "^1.0.0-rc.2",
-    "mongoose": "^5.4.2",
+    "lit-element": "2.0.1",
+    "lit-html": "^1.0.0",
+    "mongoose": "^5.4.10",
     "prismjs": "^1.15.0",
     "prouter": "^9.0.1",
-    "rxjs": "^6.3.3",
+    "rxjs": "^6.4.0",
     "showdown": "^1.9.0",
     "simplemde": "^1.11.2",
-    "typegoose": "^5.4.1",
+    "typegoose": "^5.5.0",
     "uuid": "^3.3.2"
   },
   "devDependencies": {
@@ -90,7 +91,7 @@
     "svg-inline-loader": "^0.8.0",
     "ts-loader": "^5.3.2",
     "ts-node": "^7.0.1",
-    "typescript": "^3.2.4",
+    "typescript": "^3.3.3",
     "webpack": "^4.28.4",
     "webpack-cli": "^3.2.0",
     "webpack-dev-server": "^3.1.14",
