@@ -1,6 +1,7 @@
-import { Middleware } from "@marblejs/core";
-import { tap, filter } from "rxjs/operators";
-import { Config } from "../../../config";
+import { HttpMiddlewareEffect } from '@marblejs/core';
+import { tap } from 'rxjs/operators';
+
+import { Config } from '../../../config';
 
 function getAuthorizedOrigin(origin: string | undefined): string {
   if (
@@ -17,29 +18,19 @@ function getAuthorizedOrigin(origin: string | undefined): string {
   return "";
 }
 
-export const cors$: Middleware = (req$, res) =>
+export const cors$: HttpMiddlewareEffect = (req$, res) =>
   req$.pipe(
     tap(req => {
       res.setHeader(
         "Access-Control-Allow-Origin",
         getAuthorizedOrigin(req.headers.origin as string),
       );
-      res.setHeader("Access-Control-Allow-Headers", [
-        "Content-Type",
-        "Origin",
-        "Authorization",
-      ]);
-      res.setHeader("Access-Control-Allow-Methods", [
-        "OPTIONS",
-        "GET",
-        "POST",
-        "PUT",
-        "DELETE",
-      ]);
-    }),
-    filter(req => req.method === "OPTIONS"),
-    tap(() => {
-      res.writeHead(200);
-      res.end();
+      res.setHeader("Access-Control-Allow-Headers", "Content-Type, Origin, Authorization,");
+      res.setHeader("Access-Control-Allow-Methods", "OPTIONS, GET, POST, PUT, DELETE");
+
+      if (req.method === 'OPTIONS') {
+        res.writeHead(204);
+        res.end();
+      }
     }),
   );
