@@ -1,8 +1,9 @@
-import { Effect } from "@marblejs/core";
-import { map, flatMap } from "rxjs/operators";
-import { UserDao } from "../model/user.dao";
+import { HttpEffect } from '@marblejs/core';
+import { flatMap, map } from 'rxjs/operators';
 
-export const getUsersEffect$: Effect = req$ =>
+import { UserDao } from '../model/user.dao';
+
+export const getUsersEffect$: HttpEffect = req$ =>
   req$.pipe(
     flatMap(UserDao.findAllPublic),
     map(body => ({ body })),
