@@ -1,12 +1,20 @@
-import { Effect, HttpError, HttpStatus } from "@marblejs/core";
+import { HttpEffect, HttpError, HttpStatus, use } from "@marblejs/core";
+import { requestValidator$, t } from "@marblejs/middleware-io";
 import { of, throwError } from "rxjs";
 import { catchError, map, mergeMap } from "rxjs/operators";
 
 import { neverNullable } from "../../../utils/never-nullable";
 import { ArticleDao } from "../model/article.dao";
 
-export const getArticleEffect$: Effect = req$ =>
+const validator$ = requestValidator$({
+  params: t.type({
+    id: t.string,
+  }),
+});
+
+export const getArticleEffect$: HttpEffect = req$ =>
   req$.pipe(
+    use(validator$),
     mergeMap(req =>
       of(req.params.id).pipe(
         mergeMap(ArticleDao.findById),
